requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
email-validator==2.2.0
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.1.1
Jinja2==3.1.4
MarkupSafe==2.1.5
pillow==11.0.0
python-dateutil==2.9.0.post0
pytz==2024.2
SQLAlchemy==2.0.36
Werkzeug==3.0.4
WTForms==3.1.2
psycopg2-binary==2.9.10
google-cloud-sqlcommenter==2.0.0
python-dotenv==1.0.1
google-cloud-storage==2.19.0
setuptools==75.6.0